Support caching, for faster iteration

+++ scripts/genomes/parse_genome_annotations.py
@@ -4,6 +4,7 @@
 import argparse
 import json
 import urllib.request as request
+import shutil
 import subprocess
 
 from utils import *
@@ -11,15 +12,18 @@ from utils import *
 parser = argparse.ArgumentParser(
     description=__doc__, # Use docstring at top of file for --help summary
     formatter_class=argparse.RawDescriptionHelpFormatter)
+parser.add_argument('--use_cache',
+                    help='Whether to use cache',
+                    default=True, type=bool, nargs='?')
 parser.add_argument('--output_dir',
                     help='Directory to send output data to.  Default: output/',
                     default='output/')
 args = parser.parse_args()
-credentials = args.credentials
+use_cache = args.use_cache
 output_dir = args.output_dir
 
-if os.path.exists(output_dir) is False:
-    os.mkdir(output_dir)
+if use_cache is None:
+    use_cache = False
 
 scp_species = get_species_list('organisms.tsv')
 
@@ -78,41 +82,57 @@ def get_ensembl_gtf_urls(ensembl_metadata):
     return gtf_urls
 
 def transform_ensembl_gtf(gtf_path):
-    """Produces sorted GTF and GTF index from Ensembl GTF; needed for igv.js
+    """Produce sorted GTF and GTF index from Ensembl GTF; needed for igv.js
     """
     # Example:
     # $ sort -k1,1 -k4,4n gencode.vM17.annotation.gtf > gencode.vM17.annotation.possorted.gtf
     # $ bgzip gencode.vM17.annotation.possorted.gtf
     # $ tabix -p gff gencode.vM17.annotation.possorted.gtf.gz
 
+    sorted_filename = gtf_path.replace('.gtf', '.possorted.gtf')
+    outputs = [sorted_filename + '.gz', sorted_filename + '.gz.tbi']
+    if os.path.exists(outputs[1]):
+        return outputs
+
     # sort by chromosome name, then genomic start position; needed for index
     sort_command = ('sort -k1,1 -k4,4n ' + gtf_path).split(' ')
-    sorted_filename = gtf_path.replace('.gtf', '.possorted.gtf')
     sorted_file = open(sorted_filename, 'w')
-    print('Running ' + str(sort_command))
     subprocess.call(sort_command, stdout=sorted_file)
 
     # bgzip enables requesting small indexed chunks of a gzip'd file
     bgzip_command = ('bgzip ' + sorted_filename).split(' ')
-    print('Running ' + str(bgzip_command))
     subprocess.call(bgzip_command)
 
     # tabix creates an index for the GTF file, used for getting small chunks
     tabix_command = ('tabix -p gff ' + sorted_filename + '.gz').split(' ')
-    print('Running ' + str(tabix_command))
     subprocess.call(tabix_command)
 
-def write_ensembl_gtf_products(ensembl_metadata):
-    """Download Ensembl GTFs, produce position-sorted raw and indexed GTFs
+    return outputs
+
+def transform_ensembl_gtfs(ensembl_metadata):
+    """Download raw Ensembl GTFs, write position-sorted GTF and index
     """
+    transformed_gtfs = []
+
     gtf_urls = get_ensembl_gtf_urls(ensembl_metadata)
     gtfs = batch_fetch(gtf_urls, output_dir)
+
     print('Got GTFs!  Number: ' + str(len(gtfs)))
     for gtf in gtfs:
         gtf_path = gtf[0].replace('.gz', '')
-        transform_ensembl_gtf(gtf_path)
+        transformed_gtf = transform_ensembl_gtf(gtf_path)
+        transformed_gtfs.append(transformed_gtf)
+
+if use_cache is False:
+    if os.path.exists(output_dir) is False:
+        print('Cache unavailable, starting fresh run')
+    else:
+        print('Deleting output cache from previous run')
+        shutil.rmtree(output_dir)
+
+if os.path.exists(output_dir) is False:
+    os.mkdir(output_dir)
 
 ensembl_metadata = get_ensembl_metadata()
-gtfs = get_ensembl_gtfs(ensembl_metadata)
-print('gtfs')
-print(len(gtfs))
+gtfs = get_ensembl_gtf_urls(ensembl_metadata)
+gtf_products = transform_ensembl_gtfs(ensembl_metadata)
